refactor(ui): merge `useAdminSessionsList` into `useAdminSessions`
The two hooks had overlapping responsibilities. `useAdminSessionsList`
had better code (pagination params, error handling) but the naming
was inconsistent — no other hook uses a `List` suffix. Absorb its
logic into `useAdminSessions` with `{ page, perPage }` object params
matching `useSessions`, and delete the redundant hook.

The hook-level `toDisplayError` transform is also dropped — consumers
already use `apiErrorMessage` from `@/api/errors` at render time.
