+++ ui/apps/console/src/pages/admin/Sessions.tsx
@@ -7,7 +7,7 @@ import {
 } from "@heroicons/react/24/outline";
 import { Callout } from "@shellhub/design-system/primitives";
 import { cn } from "@shellhub/design-system/cn";
-import { useAdminSessionsList } from "@/hooks/useAdminSessionsList";
+import { useAdminSessions } from "@/hooks/useAdminSessions";
 import type { Session } from "@/client";
 import PageHeader from "@/components/common/PageHeader";
 import DataTable, { type Column } from "@/components/common/DataTable";
@@ -28,10 +28,10 @@ export default function AdminSessions() {
   const { params, setPage } = usePaginatedListState<AdminSessionsParams>({
     defaults: DEFAULTS,
   });
-  const { sessions, totalCount, isLoading, error } = useAdminSessionsList(
-    params.page,
-    PER_PAGE,
-  );
+  const { sessions, totalCount, isLoading, error } = useAdminSessions({
+    page: params.page,
+    perPage: PER_PAGE,
+  });
   const navigate = useNavigate();
 
   const totalPages = Math.ceil(totalCount / PER_PAGE);
@@ -43,7 +43,12 @@ export default function AdminSessions() {
       headerClassName: "w-14",
       render: (s) => (
         <span
-          className={cn("w-2 h-2 rounded-full inline-block", s.active ? "bg-accent-green shadow-[0_0_6px_rgba(130,165,104,0.4)]" : "bg-text-muted/40")}
+          className={cn(
+            "w-2 h-2 rounded-full inline-block",
+            s.active
+              ? "bg-accent-green shadow-[0_0_6px_rgba(130,165,104,0.4)]"
+              : "bg-text-muted/40",
+          )}
         />
       ),
     },
@@ -91,7 +96,10 @@ export default function AdminSessions() {
               />
             )}
             <code
-              className={cn("text-xs font-mono", suspicious ? "text-accent-red/60" : "text-text-secondary")}
+              className={cn(
+                "text-xs font-mono",
+                suspicious ? "text-accent-red/60" : "text-text-secondary",
+              )}
             >
               {s.username}
             </code>
